Remove duplicate MatchResultView and centralize
